+++ src/main/java/kr/hhplus/be/server/application/external/dto/OrderProductData.java
@@ -3,16 +3,16 @@ package kr.hhplus.be.server.application.external.dto;
 public class OrderProductData {
 
     private Long productId;
-    private Integer amount;
-    private Integer quantity;
+    private Long amount;
+    private Long quantity;
 
-    private OrderProductData(Long productId, Integer amount, Integer quantity) {
+    private OrderProductData(Long productId, Long amount, Long quantity) {
         this.productId = productId;
         this.amount = amount;
         this.quantity = quantity;
     }
 
-    public static OrderProductData of(Long productId, Integer amount, Integer quantity) {
+    public static OrderProductData of(Long productId, Long amount, Long quantity) {
         return new OrderProductData(productId, amount, quantity);
     }
 }
