+++ src/main/java/kr/hhplus/be/server/domain/product/Product.java
@@ -13,12 +13,12 @@ public class Product {
     private Long id;
     private String name;
     private byte[] description;
-    private Integer price;
-    private Integer stock;
+    private Long price;
+    private Long stock;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
 
-    private Product(Long id, String name, byte[] description, Integer price, Integer stock) {
+    private Product(Long id, String name, byte[] description, Long price, Long stock) {
         this.id = id;
         this.name = name;
         this.description = description;
@@ -28,11 +28,11 @@ public class Product {
         this.updatedAt = LocalDateTime.now();
     }
 
-    public static Product of(Long id, String name, byte[] description, Integer price, Integer stock) {
+    public static Product of(Long id, String name, byte[] description, Long price, Long stock) {
         return new Product(id, name, description, price, stock);
     }
 
-    public void deduct(int quantity) {
+    public void deduct(Long quantity) {
         if (quantity > stock) {
             throw new ApiException(INSUFFICIENT_STOCK);
         }
