+++ src/main/java/kr/hhplus/be/server/domain/product/dto/ProductResult.java
@@ -8,10 +8,10 @@ public class ProductResult {
 
     private Long id;
     private String name;
-    private Integer price;
-    private Integer stock;
+    private Long price;
+    private Long stock;
 
-    private ProductResult(Long id, String name, Integer price, Integer stock) {
+    private ProductResult(Long id, String name, Long price, Long stock) {
         this.id = id;
         this.name = name;
         this.price = price;
