+++ src/main/java/kr/hhplus/be/server/interfaces/api/order/dto/request/OrderProductRequest.java
@@ -13,10 +13,10 @@ public class OrderProductRequest {
     private Long productId;
 
     @Min(value = 1)
-    private Integer quantity;
+    private Long quantity;
 
     @Builder
-    private OrderProductRequest(Long productId, Integer quantity) {
+    private OrderProductRequest(Long productId, Long quantity) {
         this.productId = productId;
         this.quantity = quantity;
     }
