+++ src/main/java/kr/hhplus/be/server/interfaces/api/product/dto/response/ProductResponse.java
@@ -8,10 +8,10 @@ public class ProductResponse {
 
     private Long id;
     private String name;
-    private Integer price;
-    private Integer stock;
+    private Long price;
+    private Long stock;
 
-    private ProductResponse(Long id, String name, Integer price, Integer stock) {
+    private ProductResponse(Long id, String name, Long price, Long stock) {
         this.id = id;
         this.name = name;
         this.price = price;
