+++ src/test/java/kr/hhplus/be/server/domain/point/PointServiceTest.java
@@ -30,8 +30,8 @@ class PointServiceTest {
     @DisplayName("포인트 충전을 처음하는 경우 포인트가 새로 생성되고 충전이 된다.")
     @Test
     void chargePoint_whenFirstTime() {
-        Long userId = 1L;
-        int chargeAmount = 1000;
+        long userId = 1L;
+        long chargeAmount = 1000L;
 
         given(pointRepository.findPointByUserId(userId)).willReturn(Optional.empty());
         given(pointRepository.savePoint(any(Point.class)))
@@ -49,9 +49,9 @@ class PointServiceTest {
     @DisplayName("포인트 충전이 처음이 아닌 경우 포인트가 기존에 보유 중인 포인트에 누적이 된다.")
     @Test
     void chargePoint() {
-        Long userId = 1L;
-        int initialPoint = 1000;
-        int chargeAmount = 3000;
+        long userId = 1L;
+        long initialPoint = 1000L;
+        long chargeAmount = 3000L;
 
         Point point = Point.of(2L, userId, initialPoint);
 
@@ -61,7 +61,7 @@ class PointServiceTest {
 
         Point chargedPoint = pointService.chargePoint(userId, chargeAmount);
 
-        int expectedPoint = initialPoint + chargeAmount;
+        long expectedPoint = initialPoint + chargeAmount;
         assertThat(chargedPoint.getUserId()).isEqualTo(userId);
         assertThat(chargedPoint.getBalance()).isEqualTo(expectedPoint);
 
@@ -73,7 +73,7 @@ class PointServiceTest {
     @Test
     void usePoint_whenNotExist() {
         long userId = 1L;
-        int useAmount = 1000;
+        long useAmount = 1000L;
         given(pointRepository.findPointByUserId(userId)).willReturn(Optional.empty());
 
         assertThatThrownBy(() -> pointService.usePoint(userId, useAmount))
@@ -84,9 +84,9 @@ class PointServiceTest {
     @DisplayName("포인트가 존재하는 경우 기존 보유 중인 포인트에서 사용 금액 만큼 차감된다.")
     @Test
     void usePoint() {
-        Long userId = 1L;
-        int initialPoint = 3000;
-        int useAmount = 1000;
+        long userId = 1L;
+        long initialPoint = 3000L;
+        long useAmount = 1000L;
 
         Point point = Point.of(2L, userId, initialPoint);
 
@@ -96,7 +96,7 @@ class PointServiceTest {
 
         Point chargedPoint = pointService.usePoint(userId, useAmount);
 
-        int expectedPoint = initialPoint - useAmount;
+        long expectedPoint = initialPoint - useAmount;
         assertThat(chargedPoint.getUserId()).isEqualTo(userId);
         assertThat(chargedPoint.getBalance()).isEqualTo(expectedPoint);
